+++ src/hal/src/command/render_pass.rs
@@ -12,6 +12,7 @@ use crate::{buffer, pass, pso, query};
 use crate::{Backend, DrawCount, IndexCount, InstanceCount, VertexCount, VertexOffset};
 
 /// Specifies how commands for the following renderpasses will be recorded.
+#[derive(Debug)]
 pub enum SubpassContents {
     /// Contents of the subpass will be inline in the command buffer,
     /// NOT in secondary command buffers.
@@ -27,6 +28,7 @@ pub enum SubpassContents {
 ///
 /// Where methods are undocumented, they are identical to the methods on the `RawCommandBuffer`
 /// trait with the same names.
+#[derive(Debug)]
 pub struct RenderSubpassCommon<B, C> {
     cmb: C,
     _marker: PhantomData<B>,
@@ -220,6 +222,7 @@ impl<B: Backend, C: BorrowMut<B::CommandBuffer>> RenderSubpassCommon<B, C> {
 
 /// An object that records commands into a command buffer inline, that is,
 /// without secondary command buffers.
+#[derive(Debug)]
 pub struct RenderPassInlineEncoder<'a, B: Backend>(
     RenderSubpassCommon<B, &'a mut B::CommandBuffer>,
 )
@@ -298,6 +301,7 @@ impl<'a, B: Backend> Drop for RenderPassInlineEncoder<'a, B> {
 
 /// An object that records commands into a command buffer where each command must
 /// be a call to execute a secondary command buffer.
+#[derive(Debug)]
 pub struct RenderPassSecondaryEncoder<'a, B: Backend>(&'a mut B::CommandBuffer)
 where
     B::CommandBuffer: 'a;
@@ -370,6 +374,7 @@ impl<'a, B: Backend> Drop for RenderPassSecondaryEncoder<'a, B> {
 }
 
 /// A secondary command buffer recorded entirely within a subpass.
+#[derive(Debug)]
 pub struct SubpassCommandBuffer<B: Backend, S: Shot, R = <B as Backend>::CommandBuffer>(
     RenderSubpassCommon<B, R>,
     PhantomData<S>,